+++ routes/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-csv = "1.3.0"
 log = "0.4"
 rocket = { version = "0.5.0", features=["json"] }
 rocket_cors = "0.6.0"
