+++ routes/tests/mock.rs
@@ -16,7 +16,7 @@
 #[cfg(test)]
 use maplit::hashmap;
 use scopeguard::guard;
-use shared::{registry::update_registry, reset_mock_environment, write_consumption};
+use shared::{consumption::write_consumption, registry::update_registry, reset_mock_environment};
 use std::collections::HashMap;
 use types::{ParaId, Parachain, RelayChain, RelayChain::*, WeightConsumption};
